+++ ClaysAndGlazes/App/Business Logic/ViewControllers/MaterialsList/AddMaterialViewController/ViewModel/AddMaterialViewModel.swift
@@ -25,7 +25,7 @@ class AddMaterialViewControllerViewModel: AddMaterialViewControllerViewModelType
     func addNewMaterial(type: String, quantity: String, unit: String, name: String, info: String, viewController: UIViewController) {
 
         // Create material from item parameters
-        let material = Material(type: type, name: name, quantity: Int(quantity) ?? 0, unit: unit, info: info, marked: false)
+        let material = Material(type: type, name: name, quantity: Float(quantity) ?? 0, unit: unit, info: info, marked: false)
 
         // Save material to UserDefaults
         LocalStorageService.save(object: material)
